requirements.txt
ultralytics
opencv-python
pyserial
numpy